requirements.txt
pgzero
pillow
pydantic
websocket-client
requests